Extract user helpers to senaite.jsonapi.api.users
Move is_anonymous, get_current_user, get_member_ids, get_user, and
get_user_properties out of the god-module api/__init__.py into a
focused api/users.py. The five names remain importable from
senaite.jsonapi.api via explicit re-exports at the bottom of __init__.py,
so downstream code (senaite.core, add-ons, docs, tests) needs no
change.

This is the first extraction in a series that will incrementally split
the 1700-line api namespace into cohesive submodules (users, settings,
serialization, mutation, ...) without touching the public import
surface.
