+++ src/senaite/jsonapi/api/__init__.py
@@ -1318,67 +1318,9 @@ def convert_physical_paths_to_objects(record):
     return record
 
 
-def is_anonymous():
-    """Check if the current user is authenticated or not
-
-    :returns: True if the current user is authenticated
-    :rtype: bool
-    """
-    return ploneapi.user.is_anonymous()
-
-
-def get_current_user():
-    """Get the current logged in user
-
-    :returns: Member
-    :rtype: object
-    """
-    return ploneapi.user.get_current()
-
-
-def get_member_ids():
-    """Return all member ids of the portal.
-    """
-    pm = get_tool("portal_membership")
-    member_ids = pm.listMemberIds()
-    # How can it be possible to get member ids with None?
-    return filter(lambda x: x, member_ids)
-
-
-def get_user(user_or_username=None):
-    """Return Plone User
-
-    :param user_or_username: Plone user or user id
-    :type groupname:  PloneUser/MemberData/str
-    :returns: Plone MemberData
-    :rtype: object
-    """
-    if user_or_username is None:
-        return None
-    if hasattr(user_or_username, "getUserId"):
-        return ploneapi.user.get(user_or_username.getUserId())
-    return ploneapi.user.get(userid=u.to_string(user_or_username))
-
-
-def get_user_properties(user_or_username):
-    """Return User Properties
-
-    :param user_or_username: Plone group identifier
-    :type groupname:  PloneUser/MemberData/str
-    :returns: Plone MemberData
-    :rtype: object
-    """
-    user = get_user(user_or_username)
-    if user is None:
-        return {}
-    if not callable(user.getUser):
-        return {}
-    out = {}
-    plone_user = user.getUser()
-    for sheet in plone_user.listPropertysheets():
-        ps = plone_user.getPropertysheet(sheet)
-        out.update(dict(ps.propertyItems()))
-    return out
+# User helpers live in senaite.jsonapi.api.users and are re-exported at
+# the bottom of this module so they remain importable as
+# senaite.jsonapi.api.is_anonymous / get_current_user / ...
 
 
 def find_objects(uid=None):
@@ -1707,3 +1649,13 @@ def make_batch(sequence, size=25, start=0):
     """
     # we call an adapter here to allow backwards compatibility hooks
     return IBatch(Batch(sequence, size, start))
+
+
+# Backward-compat re-exports. Kept at the bottom so any name the
+# extracted module needs from this package (e.g. helpers defined above)
+# is already bound at import time. Do not move these to the top.
+from senaite.jsonapi.api.users import is_anonymous  # noqa: E402,F401
+from senaite.jsonapi.api.users import get_current_user  # noqa: E402,F401
+from senaite.jsonapi.api.users import get_member_ids  # noqa: E402,F401
+from senaite.jsonapi.api.users import get_user  # noqa: E402,F401
+from senaite.jsonapi.api.users import get_user_properties  # noqa: E402,F401
